requirements.txt
wheel
cairo-lang
cairo-nile
pytest-cairo
tox